20180508 管理员仪器管理和管理员管理完成

+++ src/components/admin-equipmentType/admin-equipmentType.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="equipment-type">
     <el-table
+      border
       :data="tableData"
       style="width: 100%">
       <el-table-column
@@ -51,6 +52,23 @@
         <button class="btn btn-primary" slot="footer" @click.prevent.stop="addEquipmentType">确定</button>
       </Modal>
     </transition>
+    <transition name="modal">
+      <Modal :wrapperClass="'EquipmentType'" v-if="modifyEquipmentModal" @hideDetail="hideAll">
+        <div class="title" slot="header">修改分类</div>
+        <div slot="body" class="radio">
+          <el-form ref="form" :model="modifyEquipment.old" label-width="100px" :label-position="'right'">
+            <el-form-item label="分类名称：">
+              <el-input v-model="modifyEquipment.new.name" :placeholder="modifyEquipment.old.name"></el-input>
+            </el-form-item>
+            <el-form-item label="描述：">
+              <el-input type="textarea" v-model="modifyEquipment.new.description"
+                        :placeholder="modifyEquipment.old.description"></el-input>
+            </el-form-item>
+          </el-form>
+        </div>
+        <button class="btn btn-primary" slot="footer" @click.prevent.stop="modifyEquipmentType">修改</button>
+      </Modal>
+    </transition>
   </div>
 </template>
 
@@ -67,6 +85,18 @@
         addEquipment: {
           name: '',
           description: ''
+        },
+        modifyEquipmentModal: false,
+        modifyEquipment: {
+          id: 0,
+          old: {
+            name: '',
+            description: ''
+          },
+          new: {
+            name: '',
+            description: ''
+          }
         }
       }
     },
@@ -75,25 +105,68 @@
     },
     methods: {
       hideAll() {
-        this.addEquipmentModal = false
+        this.addEquipmentModal = this.modifyEquipmentModal = false
       },
       handleEdit(index, scopeRow) {
-
+        this.modifyEquipmentModal = true;
+        this.modifyEquipment.id = scopeRow.id;
+        this.modifyEquipment.old.name = scopeRow.name
+        this.modifyEquipment.old.description = scopeRow.description;
+        this.modifyEquipment.new.name =
+          this.modifyEquipment.new.description = ''
       },
-      handleDelete(index, scopeRow) {
+      modifyEquipmentType() {
+        if (!this.modifyEquipment.new.name && !this.modifyEquipment.new.description) {
+          this.$message.error(`输入不能全为空！`);
+          return;
+        }
         equipment
-          .deleteCategory({
-            id: scopeRow.id
+          .modifyCategory({
+            id: this.modifyEquipment.id,
+            name: this.modifyEquipment.new.name || this.modifyEquipment.old.name,
+            description: this.modifyEquipment.new.description || this.modifyEquipment.old.description
           })
           .then((res) => {
             if (res.ret === 200 && res.msg === 'success') {
-              this.$message.success(`分类 ${scopeRow.name} 删除成功！`)
-              this.getCategories();
+              this.$message.success(`分类 ${this.modifyEquipment.new.name} 修改成功！`)
+              this.modifyEquipment.new.name =
+                this.modifyEquipment.old.name =
+                  this.modifyEquipment.new.description =
+                    this.modifyEquipment.old.description = '';
+              this.hideAll()
+              this.getCategories()
             }
           })
           .catch((err) => {
+            console.log(err)
             this.$message.error(`[系统提醒: ${err.msg}]`);
           });
+      },
+      handleDelete(index, scopeRow) {
+        this.$confirm('此操作将永久删除该分类, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          equipment
+            .deleteCategory({
+              id: scopeRow.id
+            })
+            .then((res) => {
+              if (res.ret === 200 && res.msg === 'success') {
+                this.$message.success(`分类 ${scopeRow.name} 删除成功！`)
+                this.getCategories();
+              }
+            })
+            .catch((err) => {
+              this.$message.error(`[系统提醒: ${err.msg}]`);
+            });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });
+        });
       },
       showModal() {
         this.addEquipmentModal = true
@@ -109,6 +182,7 @@
               this.hideAll();
               this.addEquipment.name = this.addEquipment.description = ''
               this.$message.success(`分类 ${this.addEquipment.name} 创建成功！`)
+              this.getCategories();
             }
           })
           .catch((err) => {
